refactor : Club 목록 조회 쿼리 성능 개선

+++ back/Flock/src/main/java/com/Flock/domain/ClubMember/Service/ClubMemberService.java
@@ -26,8 +26,8 @@ public class ClubMemberService {
 
     private final ClubRepository clubRepository;
 
-    public Integer countClubMember(Club club) {
-        return clubMemberRepository.countClubMemberByClubAndIsMemberIsTrue(club.getId());
+    public Long countClubMember(Club club) {
+        return (long) clubMemberRepository.countClubMemberByClubAndIsMemberIsTrue(club.getId());
     }
 
     /**
